+++ BepInExFasterLoadAssetBundles/Managers/AssetBundleManager.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 namespace BepInExFasterLoadAssetBundles.Managers;
 internal class AssetBundleManager
 {
-    private string CachePath { get; }
+    public string CachePath { get; }
 
     public AssetBundleManager(string cachePath)
     {
@@ -21,79 +21,79 @@ internal class AssetBundleManager
 
     public bool TryRecompressAssetBundle(ref string path)
     {
-        try
+        var originalFileName = Path.GetFileNameWithoutExtension(path);
+        byte[] hash = HashingHelper.HashFile(path);
+
+        var metadata = Patcher.MetadataManager.FindMetadataByHash(hash);
+        if (metadata != null)
         {
-            byte[] hash;
-            using (var fileStream = File.OpenRead(path))
-            {
-                hash = HashingHelper.HashStream(fileStream);
-            }
-
-            var metadata = MetadataManager.Instance.FindMetadataByHash(hash);
-            if (metadata != null && metadata.UncompressedAssetBundleName != null)
-            {
-                var newPath = Path.Combine(CachePath, metadata.UncompressedAssetBundleName);
-                if (File.Exists(newPath))
-                {
-                    path = newPath;
-                    return true;
-                }
-
-                BepInExFasterLoadAssetBundlesPatcher.Logger.LogWarning($"Failed to find decompressed assetbundle at {newPath}. Probably it was deleted?");
-            }
-
-            if (metadata?.ShouldNotDecompress == true)
+            if (metadata.ShouldNotDecompress || metadata.UncompressedAssetBundleName == null)
             {
                 return false;
             }
 
-            metadata = new()
+            var newPath = Path.Combine(CachePath, metadata.UncompressedAssetBundleName);
+            if (File.Exists(newPath))
             {
-                OriginalAssetBundleHash = HashingHelper.HashToString(hash)
-            };
-            var outputName = Path.GetFileNameWithoutExtension(path) + '_' + metadata.GetHashCode() + ".assetbundle";
-            var outputPath = Path.Combine(CachePath, outputName);
+                BepInExFasterLoadAssetBundlesPatcher.Logger.LogDebug(
+                    $"Found uncompressed bundle {metadata.UncompressedAssetBundleName}, loading it instead of {originalFileName}");
+                path = newPath;
 
-            var op = AssetBundle.RecompressAssetBundleAsync(path, outputPath, BuildCompression.UncompressedRuntime, 0, ThreadPriority.High);
-            AsyncOperationHelper.WaitUntilOperationComplete(op);
+                metadata.LastAccessTime = DateTime.Now;
+                Patcher.MetadataManager.SaveMetadata(metadata);
 
-            if (op.result is not AssetBundleLoadResult.Success)
-            {
-                BepInExFasterLoadAssetBundlesPatcher.Logger.LogWarning($"Failed to decompress a assetbundle at {path}\n{op.humanReadableResult}");
-                return false;
+                return true;
             }
 
-            // check if unity returned the same assetbundle (means that assetbundle is already decompressed)
-            var shouldDelete = false;
-            using (var fileStream = File.OpenRead(outputPath))
-            {
-                if (hash.AsSpan().SequenceEqual(HashingHelper.HashStream(fileStream)))
-                {
-                    metadata.ShouldNotDecompress = true;
-                    MetadataManager.Instance.AddMetadata(metadata);
-
-                    shouldDelete = true;
-                }
-            }
-
-            if (shouldDelete)
-            {
-                File.Delete(outputPath);
-                return false;
-            }
-
-            path = outputPath;
-
-            metadata.UncompressedAssetBundleName = outputName;
-            MetadataManager.Instance.AddMetadata(metadata);
-
-            return true;
-        }
-        catch (Exception ex)
-        {
-            BepInExFasterLoadAssetBundlesPatcher.Logger.LogError($"Failed to decompress a assetbundle at {path}\n{ex}");
+            BepInExFasterLoadAssetBundlesPatcher.Logger.LogWarning($"Failed to find decompressed assetbundle at {newPath}. Probably it was deleted?");
         }
 
-        return false;
+        metadata = new()
+        {
+            OriginalAssetBundleHash = HashingHelper.HashToString(hash),
+            LastAccessTime = DateTime.Now,
+        };
+        var outputName = originalFileName + '_' + metadata.GetHashCode() + ".assetbundle";
+        var outputPath = Path.Combine(CachePath, outputName);
+
+        var op = AssetBundle.RecompressAssetBundleAsync(path, outputPath,
+            BuildCompression.UncompressedRuntime, 0, ThreadPriority.High);
+        AsyncOperationHelper.WaitUntilOperationComplete(op);
+
+        if (op.result is not AssetBundleLoadResult.Success)
+        {
+            BepInExFasterLoadAssetBundlesPatcher.Logger.LogWarning($"Failed to decompress a assetbundle at {path}\n{op.humanReadableResult}");
+            return false;
+        }
+
+        // check if unity returned the same assetbundle (means that assetbundle is already decompressed)
+        if (hash.AsSpan().SequenceEqual(HashingHelper.HashFile(outputPath)))
+        {
+            BepInExFasterLoadAssetBundlesPatcher.Logger.LogDebug($"Assetbundle {originalFileName} is already uncompressed, adding to ignore list");
+
+            metadata.ShouldNotDecompress = true;
+            Patcher.MetadataManager.SaveMetadata(metadata);
+
+            DeleteCachedAssetBundle(outputPath);
+            return false;
+        }
+
+        path = outputPath;
+
+        metadata.UncompressedAssetBundleName = outputName;
+        Patcher.MetadataManager.SaveMetadata(metadata);
+
+        BepInExFasterLoadAssetBundlesPatcher.Logger.LogDebug($"Loading uncompressed bundle {outputName} instead of {originalFileName}");
+
+        return true;
+    }
+
+    public void DeleteCachedAssetBundle(string path)
+    {
+        FileHelper.TryDeleteFile(path, out var fileException);
+        if (fileException != null)
+        {
+            BepInExFasterLoadAssetBundlesPatcher.Logger.LogError($"Failed to delete uncompressed assetbundle\n{fileException}");
+        }
     }
 }