+++ BepInExFasterLoadAssetBundles/Managers/MetadataManager.cs
@@ -8,16 +8,13 @@ using Newtonsoft.Json;
 namespace BepInExFasterLoadAssetBundles.Managers;
 internal class MetadataManager
 {
-    public static MetadataManager Instance { get; private set; } = null!;
-
     private readonly string m_MetadataFile;
+    private readonly object m_Lock = new();
 
     private List<Metadata> m_Metadata = null!;
 
     public MetadataManager(string metadataFile)
     {
-        Instance = this;
-
         m_MetadataFile = metadataFile;
         LoadFile();
     }
@@ -26,24 +23,35 @@ internal class MetadataManager
     {
         Span<byte> tempHash = stackalloc byte[20];
 
-        foreach (var metadata in m_Metadata)
+        lock (m_Lock)
         {
-            HashingHelper.StringToHash(metadata.OriginalAssetBundleHash, tempHash);
-            if (tempHash.SequenceEqual(hash))
+            foreach (var metadata in m_Metadata)
             {
-                return metadata;
+                HashingHelper.StringToHash(metadata.OriginalAssetBundleHash, tempHash);
+                if (tempHash.SequenceEqual(hash))
+                {
+                    return metadata;
+                }
             }
         }
 
         return null;
     }
 
-    public void AddMetadata(Metadata metadata)
+    public void SaveMetadata(Metadata metadata)
     {
-        m_Metadata.RemoveAll(m => m.OriginalAssetBundleHash == metadata.OriginalAssetBundleHash);
+        var index = m_Metadata.FindIndex(m => m.OriginalAssetBundleHash == metadata.OriginalAssetBundleHash);
 
-        m_Metadata.Add(metadata);
-        File.WriteAllText(m_MetadataFile, JsonConvert.SerializeObject(m_Metadata));
+        if (index == -1)
+        {
+            m_Metadata.Add(metadata);
+        }
+        else
+        {
+            m_Metadata[index] = metadata;
+        }
+
+        SaveFile();
     }
 
     private void LoadFile()
@@ -62,6 +70,70 @@ internal class MetadataManager
         {
             BepInExFasterLoadAssetBundlesPatcher.Logger.LogError($"Failed to deserialize metadata.json file\n{ex}");
             m_Metadata = [];
+            return;
+        }
+
+        if (UpgradeMetadata())
+        {
+            return;
+        }
+
+        DeleteOldBundles();
+    }
+
+    /// <summary>
+    /// Upgrades metadata file if some changes happened in <see cref="Metadata"/>.
+    /// </summary>
+    /// <returns><see langword="true"/> when metadata was upgraded, otherwise <see langword="false"/></returns>
+    private bool UpgradeMetadata()
+    {
+        var shouldSave = false;
+        foreach (var metadata in m_Metadata)
+        {
+            var isTimeDefault = metadata.LastAccessTime == default;
+            if (isTimeDefault)
+            {
+                metadata.LastAccessTime = DateTime.Now;
+            }
+
+            shouldSave |= isTimeDefault;
+        }
+
+        if (shouldSave)
+        {
+            SaveFile();
+        }
+
+        return shouldSave;
+    }
+
+    private void SaveFile()
+    {
+        lock (m_Lock)
+        {
+            File.WriteAllText(m_MetadataFile, JsonConvert.SerializeObject(m_Metadata));
+        }
+    }
+
+    private void DeleteOldBundles()
+    {
+        const int c_DeleteAfterDays = 3;
+
+        for (var i = m_Metadata.Count - 1; i >= 0; i--)
+        {
+            var metadata = m_Metadata[i];
+            if ((DateTime.Now - metadata.LastAccessTime).TotalDays < c_DeleteAfterDays)
+            {
+                continue;
+            }
+
+            m_Metadata.RemoveAt(i);
+            if (metadata.UncompressedAssetBundleName == null)
+            {
+                continue;
+            }
+
+            Patcher.AssetBundleManager.DeleteCachedAssetBundle(Path.Combine(Patcher.AssetBundleManager.CachePath, metadata.UncompressedAssetBundleName));
         }
     }
 }