+++ BepInExFasterLoadAssetBundles/Models/Metadata.cs
@@ -1,4 +1,6 @@
-﻿namespace BepInExFasterLoadAssetBundles.Models;
+﻿using System;
+
+namespace BepInExFasterLoadAssetBundles.Models;
 internal class Metadata
 {
     public string? UncompressedAssetBundleName { get; set; }
@@ -6,4 +8,6 @@ internal class Metadata
     public string OriginalAssetBundleHash { get; set; } = null!;
 
     public bool ShouldNotDecompress { get; set; }
+
+    public DateTime LastAccessTime { get; set; }
 }